把JudgeClient类单独抽出来
